Extract protocol row codec
